// src/app/user.interface.ts
export interface User {
    id: number;
    companies_id: number;
    name: string;
    secondName: string;
    lastName: string;
    secondLastName: string;
    dni: string;
    email: string;
    country: number;
    phone: string;
    password: string;
    was_disabled: number;
    created_at: string;
    updated_at: string;
  }