// src/app/branches/branch.interface.ts
export interface Branch {
    id: number;
    companies_id: number;
    departments_id: number;
    internal_name: string;
    internal_code: number;
    name: string;
    email: string;
    phone: string;
    city_code: string;
    location: string;
    isMain: number;
    created_at: string;
    updated_at: string;
  }